% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dftranslation.R
\name{dftranslation}
\alias{dftranslation}
\title{dftranslation}
\usage{
dftranslation(df, server, databasename)
}
\arguments{
\item{dataframe}{with two columns as validate package required.}

\item{Server}{for the supqueries and case values}

\item{Databasename}{for the supqueries and case values}
}
\value{
list of dataframes, first is a dataframe with the rules and labels for each rule. The rest is data frames of each values or output from subquery and case statement.
}
\description{
dftranslation
}
